i2c: move shared ring buffer implementation
